Move ruby mock data into seperate file
